Faster

+++ lora_attack/eval/eval.py
@@ -64,8 +64,7 @@ if __name__ == '__main__':
     if 'ft_params' in pipeline_config:
         ft_params = pipeline_config['ft_params']
         if ft_params['ft_method_type'] == 'lora':
-            model.load_adapter(peft_model_id=args.adapter_dir, device_map='cuda:0',
-                               torch_dtype=torch.float16)
+            model.load_adapter(peft_model_id=args.adapter_dir, device_map='cuda:0')
             dataset = load_dataset(eval_params['task_dataset'])
             dataset = dataset_loaders.dataset_to_loader[eval_params['task_dataset']](dataset)
             for idx, i in tqdm.tqdm(enumerate(dataset["test"])):
